WIP to split classes up, and to use namespaces and traits
