+++ examples/AUI/pane.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace WxPhpExamples\AUI;
+
+trait Pane
+{
+    
+}
